2024/operator0/src/web/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.6.2.post1
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
cryptography==43.0.3
    # via pyjwt
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi==0.112.0
    # via
    #   -r requirements.in
    #   fastapi
fastapi-cli==0.0.5
    # via
    #   fastapi
    #   fastapi-cli
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via fastapi
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
passlib==1.7.4
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.9.2
    # via fastapi
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyjwt==2.9.0
    # via
    #   -r requirements.in
    #   pyjwt
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.12
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
requests==2.31.0
    # via -r requirements.in
rich==13.9.3
    # via typer
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
starlette==0.37.2
    # via fastapi
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   rich
    #   typer
    #   uvicorn
urllib3==2.2.3
    # via requests
uvicorn==0.32.0
    # via
    #   fastapi
    #   fastapi-cli
    #   uvicorn
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
